requirements.txt
appdirs==1.4.3
asgiref==3.2.7
awsebcli==3.18.2
boto3==1.14.28
botocore==1.15.49
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cloudinary==1.24.0
colorama==0.4.3
cryptography==3.0
cycler==0.10.0
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.5
django-allauth==0.42.0
django-appconf==1.0.4
django-crispy-forms==1.9.2
django-filebrowser-no-grappelli==3.8.0
django-image-cropping==1.5.0
django-storages==1.9.1
django-tinymce==3.0.2
django-tinymce4-lite==1.8.0
docutils==0.15.2
easy-thumbnails==2.7.1
filelock==3.0.12
future==0.16.0
gunicorn==20.0.4
idna==2.7
jmespath==0.10.0
jsmin==2.2.2
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
pathspec==0.5.9
pbr==5.4.5
Pillow==7.1.2
pipenv==2018.11.26
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pypiwin32==223
PySimpleGUI==4.19.0
python-dateutil==2.8.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2019.3

requests==2.20.1
requests-oauthlib==1.3.0
s3transfer==0.3.3

social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
stevedore==1.32.0
termcolor==1.1.0
urllib3==1.24.3
virtualenv==20.0.17
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
wcwidth==0.1.9
whitenoise==5.2.0
